refactor: separate cmd registration from operations
