+++ cmd/run.go
@@ -1,0 +1,41 @@
+// Copyright Nitric Pty Ltd.
+//
+// SPDX-License-Identifier: Apache-2.0
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at:
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package cmd
+
+import (
+	"github.com/spf13/cobra"
+
+	"github.com/nitrictech/cli/pkg/command"
+	"github.com/nitrictech/cli/pkg/operations/local_run"
+)
+
+var runCmd = &cobra.Command{
+	Use:         "run",
+	Short:       "Run your project locally for development and testing",
+	Long:        `Run your project locally for development and testing`,
+	Example:     `nitric run`,
+	Annotations: map[string]string{"commonCommand": "yes"},
+	Run: func(cmd *cobra.Command, args []string) {
+		local_run.Run(cmd.Context())
+	},
+	Args: cobra.ExactArgs(0),
+}
+
+func init() {
+	runCmd.Flags().StringVarP(&envFile, "env-file", "e", "", "--env-file config/.my-env")
+	rootCmd.AddCommand(command.AddDependencyCheck(runCmd, command.Docker))
+}
