+++ pkg/operations/README.md
@@ -1,0 +1,3 @@
+# Operations
+
+Function level operations for the commands of the nitric CLI
