+++ pkg/operations/start/start.go
@@ -1,0 +1,125 @@
+package start
+
+import (
+	"context"
+	"fmt"
+	"log"
+	"os"
+	"os/signal"
+	"syscall"
+	"time"
+
+	"github.com/bep/debounce"
+	"github.com/pkg/errors"
+	"github.com/pterm/pterm"
+	"github.com/spf13/cobra"
+
+	"github.com/nitrictech/cli/pkg/dashboard"
+	"github.com/nitrictech/cli/pkg/history"
+	"github.com/nitrictech/cli/pkg/output"
+	"github.com/nitrictech/cli/pkg/project"
+	"github.com/nitrictech/cli/pkg/run"
+	"github.com/nitrictech/cli/pkg/tasklet"
+	"github.com/nitrictech/cli/pkg/utils"
+)
+
+func Run(ctx context.Context, noBrowser bool) {
+	term := make(chan os.Signal, 1)
+	signal.Notify(term, syscall.SIGTERM, syscall.SIGINT)
+
+	// Divert default log output to pterm debug
+	log.SetOutput(output.NewPtermWriter(pterm.Debug))
+	log.SetFlags(0)
+
+	config, err := project.ConfigFromProjectPath("")
+	utils.CheckErr(err)
+
+	history := &history.History{
+		ProjectDir: config.Dir,
+	}
+
+	proj := &project.Project{
+		Name:    config.Name,
+		Dir:     config.Dir,
+		History: history,
+	}
+
+	dash, err := dashboard.New(proj, map[string]string{}, noBrowser || output.CI)
+	utils.CheckErr(err)
+
+	ls := run.NewLocalServices(proj, true, dash)
+	if ls.Running() {
+		pterm.Error.Println("Only one instance of Nitric can be run locally at a time, please check that you have ended all other instances and try again")
+		os.Exit(2)
+	}
+
+	memerr := make(chan error)
+	pool := run.NewRunProcessPool()
+
+	startLocalServices := tasklet.Runner{
+		StartMsg: "Starting Local Services",
+		Runner: func(progress output.Progress) error {
+			go func(errch chan error) {
+				errch <- ls.Start(pool)
+			}(memerr)
+
+			for {
+				select {
+				case err := <-memerr:
+					// catch any early errors from Start()
+					if err != nil {
+						return err
+					}
+				default:
+				}
+				if ls.Running() {
+					break
+				}
+				progress.Busyf("Waiting for Local Services to be ready")
+				time.Sleep(time.Second)
+			}
+			return nil
+		},
+		StopMsg: "Started Local Services!",
+	}
+	tasklet.MustRun(startLocalServices, tasklet.Opts{
+		Signal: term,
+	})
+
+	pterm.DefaultBasicText.Println("Local running, use ctrl-C to stop")
+
+	stackState := run.NewStackState(proj)
+
+	area, _ := pterm.DefaultArea.Start()
+	// Create a debouncer for the refresh and remove locking
+	debounced := debounce.New(500 * time.Millisecond)
+
+	// React to worker pool state and update services table
+	pool.Listen(func(we run.WorkerEvent) {
+		debounced(func() {
+			err := ls.Refresh()
+			if err != nil {
+				cobra.CheckErr(err)
+			}
+
+			stackState.Update(pool, ls)
+
+			area.Update(stackState.Tables(9001, *ls.GetDashPort()))
+
+			for _, warning := range stackState.Warnings() {
+				pterm.Warning.Println(warning)
+			}
+		})
+	})
+
+	select {
+	case membraneError := <-memerr:
+		fmt.Println(errors.WithMessage(membraneError, "membrane error, exiting"))
+	case <-term:
+		fmt.Println("Shutting down services - exiting")
+	}
+
+	_ = area.Stop()
+	// Stop the membrane
+	utils.CheckErr(ls.Stop())
+}
